scripts/requirements.txt
black
click
pyyaml
